Stencils

Stencils.AbstractKernelStencil — Type

AbstractKernelStencil <: Stencil

Abstract supertype for kernel stencils.

These can wrap any other stencil object, and include a kernel of the same length and positions as the stencil.

Stencils.AbstractPositionalStencil — Type

AbstractPositionalStencil <: Stencil

Positional stencils are tuples of coordinates that are specified in relation to the central point of the current cell. They can be any arbitrary shape or size, but should be listed in column-major order for performance.

Stencils.AbstractStencilArray — Type

AbstractStencilArray <: StaticArray

Supertype for arrays with a Stencil, a BoundaryCondition, and Padding.

Stencils.AbstractSwitchingStencilArray — Type

AbstractSwitchingStencilArray

Abstract supertype for AbstractStencilArray that wrap two arrays that switch places with each broadcast.

Stencils.AngledCross — Type

AngledCross <: Stencil

AngledCross(; radius=1, ndims=2)
AngledCross(radius, ndims)
AngledCross{R,N}()
AngledCross{R,N}()

A neighboorhood where offsets of zero on at least N-1 axes are included in the neighborhoods

Using R and N type parameters removes runtime cost of generating the stencil, compated to passing arguments/keywords.

Stencils.BackSlash — Type

BackSlash <: Stencil

BackSlash(; radius=1, ndims=2)
BackSlash(radius, ndims)
BackSlash{R,N}()
BackSlash{R,N}()

A neighboorhood where offsets of zero on at least N-1 axes are included in the neighborhoods

Using R and N type parameters removes runtime cost of generating the stencil, compated to passing arguments/keywords.

Stencils.BoundaryCondition — Type

BoundaryCondition

Abstract supertype for flags that specify the boundary conditions. These determine what happens when a stencil extends outside of the grid.

Stencils.Circle — Type

Circle <: Stencil

Circle(; radius=1, ndims=2)
Circle(radius, ndims)
Circle{R,N}()
Circle{R,N}()

A circular stencil

Using R and N type parameters removes runtime cost of generating the stencil, compated to passing arguments/keywords.

Stencils.Conditional — Type

Conditional <: Padding

Padding that doesn't change the array size, but checks getindex for out-of-bounds indexing, and inserts padval with Remove or values from the other side of the array with Wrap.

Stencils.Cross — Type

Cross <: Stencil

Cross(; radius=1, ndims=2)
Cross(radius, ndims)
Cross{R,N}()
Cross{R,N}()

A cross-shaped neighboorhood where offsets of zero on at least N-1 axes are included in the neighborhoods

Using R and N type parameters removes runtime cost of generating the stencil, compated to passing arguments/keywords.

Stencils.ForwardSlash — Type

ForwardSlash <: Stencil

ForwardSlash(; radius=1, ndims=2)
ForwardSlash(radius, ndims)
ForwardSlash{R,N}()
ForwardSlash{R,N}()

A neighboorhood where offsets of zero on at least N-1 axes are included in the neighborhoods

Using R and N type parameters removes runtime cost of generating the stencil, compated to passing arguments/keywords.

Stencils.Halo — Type

Halo{X} <: Padding

Padding that uses an in-memory halo around the array so that parts of a stencil that go off the edge of the array can index directly into it without a bounds check or any conditional. This has the benefit of possibly better performance during window broadcasts, but some downsides.

In :out mode, a whole new array is alocated, larger than the original. This may not be worth doing unless you are using it multiple times. with :in mode, the outside edge of the array is used as padding. This may be more accurate as there are no boundary effects from using a padding value.:w

Example

halo_in = Halo(:in)
halo_out = Halo(:out)

Stencils.Horizontal — Type

Horizontal <: Stencil

Horizontal(; radius=1, ndims=2)
Horizontal(radius, ndims)
Horizontal{R,N}()
Horizontal{R,N}()

A horizontal bar

Using R and N type parameters removes runtime cost of generating the stencil, compated to passing arguments/keywords.

Stencils.Kernel — Type

Kernel <: AbstractKernelStencil

Kernel(stencil, kernel)

Wrap any other stencil object, and includes a kernel of the same length and positions as the stencil.

Stencils.Layered — Type

Layered <: Abstract

Layered(layers::...)

Tuple or NamedTuple of stencils that can be used together.

neighbors for Layered returns a tuple of iterators for each stencil layer.

Stencils.Moore — Type

Moore <: Stencil

Moore(; radius=1, ndims=2)
Moore(radius, ndims)
Moore{R,N}()
Moore{R,N}()

Moore stencils define the stencil as all cells within a horizontal or vertical distance of the central cell. The central cell is omitted.

Using R and N type parameters removes runtime cost of generating the stencil, compated to passing arguments/keywords.

Stencils.Padding — Type

Padding

Abstract supertype for padding modes, e.g. Conditional and Halo.

Stencils.Positional — Type

Positional <: AbstractPositionalStencil

Positional(coord::Tuple{Vararg{Int}}...)
Positional(offsets::Tuple{Tuple{Vararg{Int}}})
Positional{O}()

Stencils that can take arbitrary shapes by specifying each coordinate, as Tuple{Int,Int} of the row/column distance (positive and negative) from the central point.

The stencil radius is calculated from the most distant coordinate. For simplicity the window read from the main grid is a square with sides 2r + 1 around the central point.

The dimensionality N of the stencil is taken from the length of the first coordinate, e.g. 1, 2 or 3.

Example radius R = 1:

N = 1   N = 2

 ▄▄      ▀▄
          ▀

Example radius R = 2:

N = 1   N = 2

         ▄▄
 ▀ ▀▀   ▀███
           ▀

Using the O parameter e.g. Positional{((1, 2), (1, 1))}() removes any runtime cost of generating the stencil.

Stencils.Remove — Type

Remove <: BoundaryCondition

Remove()

BoundaryCondition flag that specifies to assign padval to cells that overflow grid boundaries.

Stencils.Stencil — Type

Stencil <: StaticVector

Stencils define a pattern of neighboring cells around the current cell. They reduce the dimensions of the neighborhood values into a StaticVector of neighbor values.

Stencils objects are updated to contain the neighbors of a location. This is so that user functions can be passed a single object from whitch they can retreive neighbors, distances to neighbors and other information, rather than having this in multiple objects.

Stencils also provide a range of compile-time utility funcitons like distances and offsets.

Stencils.StencilArray — Type

StencilArray <: AbstractStencilArray

An array with a Stencil and a BoundaryCondition, and Padding.

For most uses a StencilArray works exactly the same as a regular array.

Except it can be indexed at any point with stencil to return a filled Stencil object, or neighbors to return an SVector of neighbors.

Example

``` using Stencils A = StencilArray((1:10) * (10:20)'; stencil=Moore(2), boundary=Wrap()) A .*= 2 # Broadcast works as usual hood = stencil(A, 5, 10)

ouput

Moore{1, 2, 8, StaticArraysCore.SVector{8, Int64}} █▀█ ▀▀▀ with neighbors: 8-element StaticArraysCore.SVector{8, Int64} with indices SOneTo(8): 144 180 216 152 228 160 200 240

Stencils.SwitchingStencilArray — Type

SwitchingStencilArray <: AbstractSwitchingStencilArray

An AbstractArray with a Stencil, a BoundaryCondition, Padding, and two array layers that are switched with each broadcast_stencil operation.

The use case for this operation is in simulations where stencil operations are repeatedly run over the same data, or where a filter (such as a blur) needs to be applied many times.

For most uses a SwitchingStencilArray works exactly the same as a regular array - the dest array can be safely ignored.

Example

``` using Stencils A = SwitchingStencilArray((1:10) * (10:20)'; stencil=Moore(2), boundary=Wrap()) A .*= 2 # Broadcast works as usual hood = stencil(A, 5, 10)

ouput

Moore{1, 2, 8, StaticArraysCore.SVector{8, Int64}} █▀█ ▀▀▀ with neighbors: 8-element StaticArraysCore.SVector{8, Int64} with indices SOneTo(8): 144 180 216 152 228 160 200 240

Stencils.Vertical — Type

Vertical <: Stencil

Vertical(; radius=1, ndims=2)
Vertical(radius, ndims)
Vertical{R,N}()
Vertical{R,N}()

A vertical bar

Using R and N type parameters removes runtime cost of generating the stencil, compated to passing arguments/keywords.

Stencils.VonNeumann — Type

VonNeumann <: Stencil

VonNeumann(; radius=1, ndims=2)
VonNeumann(radius, ndims)
VonNeumann{R,N}()
VonNeumann{R,N}()

Diamond-shaped neighborhood (in 2 dimwnsions), without the central cell In 1 dimension it is identical to Moore.

Using R and N type parameters removes runtime cost of generating the stencil, compated to passing arguments/keywords.

Stencils.Window — Type

Window <: Stencil

Window(; radius=1, ndims=2)
Window(radius, ndims)
Window{R,N}()
Window{R,N}()

A neighboorhood of radius R that includes the central cell.

Using R and N type parameters removes runtime cost of generating the stencil, compated to passing arguments/keywords.

Stencils.Wrap — Type

Wrap <: BoundaryCondition

Wrap()

BoundaryCondition flag to wrap cordinates that boundary boundaries back to the opposite side of the grid.

Stencils.boundary — Method

boundary(A::AbstractStencilArray)

Get the BoundaryCondition object from an AbstractStencilArray.

Stencils.diameter — Method

diameter(rule) -> Int

The diameter of a stencil is 2r + 1 where r is the radius.

Stencils.distance_zones — Method

distance_zones(hood::Stencil)

List all distance zones as a Tuple

Stencils.distances — Method

distances(hood::Stencil)

Get the center-to-center distance of each stencil position from the central cell, so that horizontally or vertically adjacent cells have a distance of 1.0, and a diagonally adjacent cell has a distance of sqrt(2.0).

Values are calculated at compile time, so distances can be used with little overhead.

Stencils.indices — Function

indices(x::Stencil, I::Tuple) -> iterable

Returns an indexable iterable of Tuple indices of each neighbor in the main array.

Stencils.inner_axes — Method

inner_axes(A, radius)

Remove padding of radius from axes of A, returning a Tuple of UnitRange. radius can be a Stencil, an Int, or a tuple of tuples, e.g. for 2d it could be: ((1, 2), (2, 1))::Tuple{Tuple{Int,Int},Tuple{Int,Int}}.

Stencils.inner_view — Method

inner_view(A, radius)

Remove padding of radius from array A, returning a view of A.

radius can be a Stencil, an Int, or a tuple of tuples, e.g. for 2d it could be: ((1, 2), (2, 1))::Tuple{Tuple{Int,Int},Tuple{Int,Int}}.

Stencils.kernel — Function

kernel(hood::AbstractKernelStencil) => iterable

Returns the kernel object, an array or iterable matching the length of the stencil.

Stencils.kernelproduct — Function

kernelproduct(hood::AbstractKernelStencil)
kernelproduct(hood::Stencil, kernel)

Returns the vector dot product of the stencil and the kernel, although differing from dot in that the dot product is not taken iteratively for members of the stencil - they are treated as scalars.

Stencils.kernelproduct — Method

kernelproduct(hood::AbstractKernelStencil)
kernelproduct(hood::Stencil, kernel)

Take the vector dot produce of the stencil and the kernel, without recursion into the values of either. Essentially Base.dot without recursive calls on the contents, as these are rarely what is intended.

Stencils.mapstencil! — Method

mapstencil!(f, dest, source::StencilArray, args...)
mapstencil!(f, A::SwitchingStencilArray, args...)

Stencil broadcast where f is passed each stencil of src, writing the result of f to dest.

For SwitchingStencilArray the internal source and dest arrays are used, returning a switched version of the array.

dest must either be smaller than src by the stencil radius on all sides, or be the same size, in which case it is assumed to also be padded.

Stencils.mapstencil — Method

mapstencil(f, hood::Stencil, args...)

Simple stencil application, where f is passed each stencil in A, returning a new array.

The result is smaller than A on all sides, by the stencil radius.

Stencils.neighbor_getindex — Method

neighbor_getindex(A::AbstractStencilArray, I::CartesianIndex)

Get an array value from the stencil neighborhood.

This method handles boundary conditions.

Stencils.neighbors — Function

neighbors(x::Stencil) -> iterable

Returns an indexable iterator for all cells in the stencil, either a Tuple of values or a range.

Custom Stencils must define this method.

Stencils.neighbors — Method

neighbors(hood::Stencil, A::AbstractArray, I) => SArray

Get a single stencil from an array, as a Tuple, checking bounds.

Stencils.offsets — Function

offsets(x) -> iterable

Returns an indexable iterable over all cells in the stencil, containing Tuples of the offset from the central cell.

Custom Stencils must define this method.

Stencils.outer_axes — Method

outer_axes(A, hood::Stencil{R})
outer_axes(A, radius::Int)

Add padding to axes of array A, returning a Tuple of UnitRange. radius can be a Stencil, an Int, or a tuple of tuples, e.g. for 2d it could be: ((1, 2), (2, 1))::Tuple{Tuple{Int,Int},Tuple{Int,Int}}.

Stencils.pad_array — Method

outer_array(A, radius; [padval])

Add padding of radius to array A, redurning a new array.

radius can be a Stencil, an Int, or a tuple of tuples, e.g. for 2d it could be: ((1, 2), (2, 1))::Tuple{Tuple{Int,Int},Tuple{Int,Int}}.

padval defaults to zero(eltype(A)).

Stencils.padding — Method

padding(A::AbstractStencilArray)

Get the Padding object from an AbstractStencilArray.

Stencils.radius — Function

radius(stencil) -> Int

Return the radius of a stencil.

Stencils.setneighbors — Function

setneighbors(x::Stencil, neighbors::StaticArray)

Update the eighbors of a Stencil, returning and identical object with new values.

Stencils.stencil — Function

stencil(A::AbstractStencilArray, I::Tuple)

Get a Stencil with neighbors updated for indices I.

Stencils.unsafe_neighbors — Method

unsafe_neighbors(hood::Stencil, A::AbstractArray, I) => SArray

Get a single stencil from an array, as a Tuple, without checking bounds of I.

Stencils.unsafe_update_stencil — Method

unsafe_update_stencil(x, A::AbstractArray, I) => Stencil

Set the neighbors of a stencil to values from the array A around index I, without checking bounds of I.

Stencils.update_stencil — Method

update_stencil(x, A::AbstractArray, I) => Stencil

Set the neighbors of a stencil to values from the array A around index I. Bounds checks will reduce performance, aim to use unsafe_setneighbors directly.