Stencils

Main functions

Stencils.mapstencil — Function

mapstencil(f, A::StencilArray, args::AbstractArray...)
mapstencil(f, stencil::Stencil, A::AbstractArray, args::AbstractArray...)

Stencil mapping where f is passed a Stencil centered at each index in A, followed by the values from args at each stencil center.

The result is returned as a new array.

Stencils.mapstencil! — Function

mapstencil!(f, dest::AbstractArray, source::StencilArray, args::AbstractArray...)
mapstencil!(f, A::SwitchingStencilArray, args::AbstractArray...)

Stencil mapping where f is passed a stencil centered at each index in src, followed by the values from args at each stencil center. The result of f is written to dest.

For SwitchingStencilArray the internal source and dest arrays are used, returning a switched version of the array.

dest must either be smaller than src by the stencil radius on all sides, or be the same size, in which case it is assumed to also be padded.

Stencils

Stencils.Stencil — Type

Stencil <: StaticVector

Stencils define a pattern of neighboring cells around the current cell. They reduce the structure and dimensions of the neighborhood into a StaticVector of values.

Stencil objects are updated to contain the neighbors for an array index.

This design is so that user functions can be passed a single object from which they can retrieve neighbors, offsets, distances to neighbors and other information.

Stencils also provide a range of compile-time utility funcitons like distances and offsets.

Stencils.NamedStencil — Type

NamedStencil <: AbstractStencil

NamedStencil(; kw...)
NamedStencil(values::NamedTuple)
NamedStencil{Keys}(values)

A named stencil that can take arbitrary shapes where each offset position is named. This can make stencil code easier to read by removing magic numbers.

Example

julia> using Stencils

julia> ns = NamedStencil(; west=(0, -1), north=(1, 0), south=(-1, 0), east=(0, 1)) 
NamedStencil{(:west, :north, :south, :east), ((0, -1), (1, 0), (-1, 0), (0, 1)), 1, 2, 4, Nothing}
▄▀▄
 ▀ 

julia> A = StencilArray((1:10) * (1:10)', ns);

julia> stencil(A, (5, 5)).east # we can access values by name
30

julia> mapstencil(s -> s.east + s.west, A); # and use them in `mapstencil` functions

The stencil radius is calculated from the most distant coordinate, and the dimensionality N of the stencil is taken from the length of the first coordinate, e.g. 1, 2 or 3.

Stencils.AngledCross — Type

AngledCross <: Stencil

AngledCross(; radius=1, ndims=2)
AngledCross(radius, ndims)
AngledCross{R,N}()

A neighboorhood where all diagonals are included, for 2:N dimensions

Stencils.BackSlash — Type

BackSlash <: Stencil

BackSlash(; radius=1, ndims=2)
BackSlash(radius, ndims)
BackSlash{R,N}()

A neighboorhood along the 'backwards' diagonal. Contains 2R+1 neighbors for for 2:N dimensions

Stencils.Circle — Type

Circle <: Stencil

Circle(; radius=1, ndims=2)
Circle(radius, ndims)
Circle{R,N}()

A circular or spherical stencil

Stencils.Cross — Type

Cross <: Stencil

Cross(; radius=1, ndims=2)
Cross(radius, ndims)
Cross{R,N}()

A cross-shaped neighboorhood where positions with offsets of 0 on least N-1 axes are included

Stencils.Diamond — Type

Diamond <: Stencil

Diamond(; radius=1, ndims=2)
Diamond(radius, ndims)
Diamond{R,N}()

A diamond or regular octahedron

Stencils.ForwardSlash — Type

ForwardSlash <: Stencil

ForwardSlash(; radius=1, ndims=2)
ForwardSlash(radius, ndims)
ForwardSlash{R,N}()

A neighboorhood where only 'forward' diagonals are included. Contains 2R+1 neighbors for 2:N dimensions

Stencils.Kernel — Type

Kernel <: AbstractKernelStencil

Kernel(stencil, kernel)

Wrap any other stencil object, and includes a kernel of the same length and positions as the stencil.

Stencils.Horizontal — Type

Horizontal <: Stencil

Horizontal(; radius=1, ndims=2)
Horizontal(radius, ndims)
Horizontal{R,N}()

A horizontal bar or plane

Stencils.Layered — Type

Layered <: Abstract

Layered(layers::...)

Tuple or NamedTuple of stencils that can be used together.

neighbors for Layered returns a tuple of iterators for each stencil layer.

Stencils.Moore — Type

Moore <: Stencil

Moore(; radius=1, ndims=2)
Moore(radius, ndims)
Moore{R,N}()

Moore stencils define the stencil as all cells within a horizontal or vertical distance of the central cell. The central cell is omitted.

Stencils.Positional — Type

Positional <: AbstractPositionalStencil

Positional(offsets::Tuple{Vararg{Int}}...)
Positional(offsets::Tuple{Tuple{Vararg{Int}}})
Positional{O}()

Stencils that can take arbitrary shapes by specifying each coordinate, as Tuple{Int,Int} of the row/column distance (positive and negative) from the central point.

The stencil radius is calculated from the most distant coordinate, and the dimensionality N of the stencil is taken from the length of the first coordinate, e.g. 1, 2 or 3.

See NamedStencil for a similar stencil with named offsets.

Example

julia> p = Positional((0, -1), (2, 1), (-1, 1), (0, 1)) 
Positional{((0, -1), (2, 1), (-1, 1), (0, 1)), 2, 2, 4, Nothing}
   ▄ 
 ▀ ▀ 
   ▀ 

Stencils.Vertical — Type

Vertical <: Stencil

Vertical(; radius=1, ndims=2)
Vertical(radius, ndims)
Vertical{R,N}()

A vertical bar or plane

Stencils.VonNeumann — Type

VonNeumann <: Stencil

VonNeumann(; radius=1, ndims=2)
VonNeumann(radius, ndims)
VonNeumann{R,N}()

Diamond-shaped neighborhood (in 2 dimwnsions), without the central cell In 1 dimension it is identical to Moore.

Stencils.Window — Type

Window <: Stencil

Window(; radius=1, ndims=2)
Window(radius, ndims)
Window{R,N}()

A neighboorhood of radius R that includes the central cell.

Stencil functions

These can be called on any <: Stencil object, and Layered.

Stencils.stencil — Function

stencil(x) -> Stencil

Returns a stencil object.

Stencils.neighbors — Function

neighbors(x::Stencil) -> iterable

Returns a basic SVector of all cells in the stencil.

Stencils.offsets — Function

offsets(x)

Return an SVector of NTuple{N,Int}, containing all positions in the stencil as offsets from the central cell.

Custom Stencils must define this method.

Stencils.indices — Function

indices(x::Stencil, I::Union{Tuple,CartesianIndex})

Returns an SVector of CartesianIndices for each neighbor around I.

Stencils.distances — Function

distances(hood::Stencil)

Returns an SVector of center-to-center distance of each stencil position from the central cell, so that horizontally or vertically adjacent cells have a distance of 1.0, and a diagonally adjacent cell has a distance of sqrt(2.0).

Values are calculated at compile time, so distances can be used with little overhead.

Stencils.radius — Function

radius(stencil) -> Int

Return the radius of a stencil.

Stencils.diameter — Function

diameter(rule) -> Int

The diameter of a stencil is 2r + 1 where r is the radius.

Stencils.distance_zones — Function

distance_zones(hood::Stencil)

Returns an SVector of Int distance zones for each offset in the Stencil.

Stencils.kernel — Function

kernel(hood::AbstractKernelStencil) => iterable

Returns the kernel object, an array or iterable matching the length of the stencil.

Stencils.kernelproduct — Function

kernelproduct(hood::AbstractKernelStencil)
kernelproduct(hood::Stencil, kernel)

Returns the vector dot product of the stencil and the kernel, although differing from dot in that it is not taken iteratively for members of the stencil - they are treated as scalars.

Stencils.unsafe_neighbors — Function

unsafe_neighbors([hood::Stencil,] A::AbstractStencilArray, I::CartesianIndex) => SArray

Get stencil neighbors from A around center I as an SVector, without checking bounds of I.

Stencils.unsafe_stencil — Function

unsafe_stencil(x, A::AbstractArray, I) => Stencil

Update the neighbors of a stencil to values from the array A around index I, without checking bounds of I. Bounds checking of neighbors still occurs, but with the assumption that I is inbounds.

Stencils.rebuild — Function

rebuild(x::Stencil, neighbors::StaticArray)

Rebuild a Stencil, returning an stencil of the same size and shape, with new neighbor values.

Stencils.getneighbor — Function

getneighbor(A::AbstractStencilArray, I::CartesianIndex)

Get an array value from a stencil neighborhood.

This method handles boundary conditions.

Boundary Conditions

Stencils.BoundaryCondition — Type

BoundaryCondition

Abstract supertype for flags that specify the boundary conditions. These determine what happens when a stencil extends outside of the grid.

Stencils.Remove — Type

Remove <: BoundaryCondition

Remove()

BoundaryCondition flag that specifies to assign padval to cells that overflow grid boundaries.

Stencils.Use — Type

Use <: BoundaryCondition

Use()

BoundaryCondition flag that specifies to use the existing padding, which is only possible when Halo{:in} is used for padding.

Stencils.Wrap — Type

Wrap <: BoundaryCondition

Wrap()

BoundaryCondition flag to wrap cordinates that boundary boundaries back to the opposite side of the grid.

Padding

Stencils.Padding — Type

Padding

Abstract supertype for padding modes, e.g. Conditional and Halo.

Stencils.Conditional — Type

Conditional <: Padding

Padding that doesn't change the array size, but checks getindex for out-of-bounds indexing, and inserts padval with Remove or values from the other side of the array with Wrap.

Stencils.Halo — Type

Halo{X} <: Padding

Padding that uses an in-memory halo around the array so that parts of a stencil that go off the edge of the array can index directly into it without a bounds check or any conditional. This has the benefit of possibly better performance during window broadcasts, but some downsides.

In :out mode, a whole new array is alocated, larger than the original. This may not be worth doing unless you are using it multiple times. with :in mode, the outside edge of the array is used as padding. This may be more accurate as there are no boundary effects from using a padding value.:w

Example

halo_in = Halo(:in)
halo_out = Halo(:out)

Stencil Arrays

Stencils.AbstractStencilArray — Type

AbstractStencilArray <: StaticArray

Supertype for arrays with a Stencil, a BoundaryCondition, and Padding.

Stencils.StencilArray — Type

StencilArray <: AbstractStencilArray

An array with a Stencil and a BoundaryCondition, and Padding.

For most uses a StencilArray works exactly the same as a regular array.

Except it can be indexed at any point with stencil to return a filled Stencil object, or neighbors to return an SVector of neighbors.

Example

using Stencils, Statistics
sa = StencilArray((1:10) * (10:20)', Moore(1); boundary=Wrap())
sa .*= 2 # Broadcast works as usual
means = mapstencil(mean, sa) # mapstencil works
stencil(sa, 5, 6) # manually reading a stencil works too

# output

Moore{1, 2, 8, Int64}
█▀█
▀▀▀

with neighbors:
8-element StaticArraysCore.SVector{8, Int64} with indices SOneTo(8):
 112
 140
 168
 120
 180
 128
 160
 192

Stencils.AbstractSwitchingStencilArray — Type

AbstractSwitchingStencilArray

Abstract supertype for AbstractStencilArray that wrap two arrays that switch places with each broadcast.

Stencils.SwitchingStencilArray — Type

SwitchingStencilArray <: AbstractSwitchingStencilArray

An AbstractArray with a Stencil, a BoundaryCondition, Padding, and two array layers that are switched with each broadcast_stencil operation.

The use case for this operation is in simulations where stencil operations are repeatedly run over the same data, or where a filter (such as a blur) needs to be applied many times.

For most uses a SwitchingStencilArray works exactly the same as a regular array - the dest array can be safely ignored.

However, when using mapstencil! you need to use the output, not the original array. Switching does not happen in-place, but as a new returned array.

Example

using Stencils, Statistics

sa = SwitchingStencilArray(rand(10, 10), Moore(1); boundary=Wrap())
sa .*= 2 # Broadcast works as usual
mapstencil(mean, sa) # As does runing `mapstencils
hood = stencil(sa, 5, 10) # And retreiving a stencil
# But we can also run it in-place, here doing 10 iterations of mean blur:
# Note: if you dont assign new variable with `A =`, the array will
# not switch and will not be blurred.
let sa = sa
    for i in 1:10
        sa = mapstencil!(mean, sa)
    end
end
# output

Methods on stencil arrays:

Stencils.boundary — Function

boundary(A::AbstractStencilArray)

Get the BoundaryCondition object from an AbstractStencilArray.

Stencils.padding — Function

padding(A::AbstractStencilArray)

Get the Padding object from an AbstractStencilArray.

Stencils.switch — Function

switch(A::SwitchingStencilArray)

Swap the source and dest of a SwitchingStencilArray.